Before transfer to client

+++ utils/checkAll.js
@@ -1,4 +1,4 @@
-import { addDays, format, parse, nextMonday, nextTuesday, nextWednesday, nextThursday, nextFriday, nextSaturday, nextSunday } from 'date-fns';
+import { addDays, format, parse, nextMonday, nextTuesday, nextWednesday, nextThursday, nextFriday, nextSaturday, nextSunday, nextDay } from 'date-fns';
 import Database from "better-sqlite3";
 import { fr, pl } from 'date-fns/locale';
 
@@ -29,6 +29,7 @@ export function checkAll() {
     if (recurrences.length > 0) {
         for (const recurrence of recurrences) {
             const nextDate = parse(recurrence.next_day, 'dd/MM/yyyy', new Date(), { locale: fr });
+            const startDateParsed = parse(recurrence.start_date, 'dd/MM/yyyy', new Date(), { locale: fr });
             const plannings = db.prepare('SELECT * FROM planning WHERE recurrence_id = ?').all(recurrence.id);
             const last = plannings[plannings.length - 1];
             const now = new Date();
@@ -49,11 +50,17 @@ export function checkAll() {
                 if (frequency.length === 1) {
                     const nextDay = nextWeekDay[frequency[0]];
                     newDate = format(nextDay(now), 'dd/MM/yyyy', { locale: fr });
-                    console.log(newDate, frequency, recurrence.id);
+                    let startDateLoop = now;
+                    let alldates = [];
+                    for (let i = 0; i < 4; i++) {
+                        const potentialNextDate = nextWeekDay[frequency[0]](startDateLoop);
+                        alldates.push(potentialNextDate);
+                        startDateLoop = potentialNextDate;
+                    }
 
                     plannings.map(planning => {
                         const parsedDate = parse(planning.date, 'dd/MM/yyyy', new Date(), { locale: fr });
-                        if (parsedDate > nextDay(now)) {
+                        if (parsedDate > nextDay(now) && !alldates.some(date => date === parsedDate)) {
                             db.prepare('DELETE FROM planning WHERE id = ?').run(planning.id);
                         }
                     });
@@ -62,15 +69,25 @@ export function checkAll() {
                     if (!isPlanned) {
                         db.prepare('INSERT INTO planning (driver_id, date, client_name, start_time, return_time, note, destination, long_distance, recurrence_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)').run(0, format(nextDay(now), 'dd/MM/yyyy', { locale: fr }), last.client_name, last.start_time, last.return_time, last.note, last.destination, last.long_distance, recurrence.id);
                     }
+                    alldates.map(date => {
+                        const formatedDate = format(date, 'dd/MM/yyyy', { locale: fr });
+                        if (formatedDate !== newDate && !plannings.some(planning => planning.date === formatedDate)) {
+                            db.prepare('INSERT INTO planning (driver_id, date, client_name, start_time, return_time, note, destination, long_distance, recurrence_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)').run(0, formatedDate, last.client_name, last.start_time, last.return_time, last.note, last.destination, last.long_distance, recurrence.id);
+                        }
+                    });
                 } else if (frequency.length > 1) {
                     let nextDates = [];
                     let allNextDates = [];
                     for (const day of frequency) {
-                        const potentialNextDate = nextWeekDay[day](now);
-                        allNextDates.push(potentialNextDate);
-                        const compareDate = potentialNextDate - now;
-                        const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
-                        nextDates.push({ [day]: difference });
+                        let currentDate = now;
+                        for (let i = 0; i < 30; i++) {
+                            const potentialNextDate = nextWeekDay[day](currentDate);
+                            allNextDates.push(potentialNextDate);
+                            const compareDate = potentialNextDate - now;
+                            const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
+                            nextDates.push({ [day]: difference });
+                            currentDate = potentialNextDate;
+                        }
                     }
                     nextDates.sort((a, b) => a[Object.keys(a)[0]] - b[Object.keys(b)[0]]);
                     let minKey = Object.keys(nextDates[0])[0];
@@ -99,8 +116,7 @@ export function checkAll() {
                 const isPlanned = plannings.find(planning => planning.date === format(nextDate, 'dd/MM/yyyy', { locale: fr }));
                 const parsedStartDate = parse(recurrence.start_date, 'dd/MM/yyyy', new Date(), { locale: fr });
                 if (frequency.length === 1) {
-                    console.log('one recurrence', frequency, recurrence.id);
-                    const nextDay = nextWeekDay[frequency[0]];
+                    const nextDay = nextWeekDay[frequency[0]]; // Utiliser plus tard dans l'insertion dans la base de donnée
                     if (!isPlanned) {
                         try {
                             db.prepare('INSERT INTO planning (driver_id, date, client_name, start_time, return_time, note, destination, long_distance, recurrence_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)').run(0, format(nextDay(parsedStartDate), 'dd/MM/yyyy'), plannings[0].client_name, plannings[0].start_time, plannings[0].return_time, plannings[0].note, plannings[0].destination, plannings[0].long_distance, recurrence.id);
@@ -112,22 +128,38 @@ export function checkAll() {
                             }
                         }
                     }
+                    let startDateLoop = parsedStartDate;
+                    let alldates = [];
+                    for (let i = 0; i < 4; i++) {
+                        const potentialNextDate = nextWeekDay[frequency[0]](startDateLoop);
+                        alldates.push(potentialNextDate);
+                        startDateLoop = potentialNextDate;
+                    }
                     plannings.map(planning => {
                         const formated = format(nextDay(parsedStartDate), 'dd/MM/yyyy', { locale: fr });
-                        if (planning.date > recurrence.start_date && planning.date !== formated) {
+                        if (planning.date > recurrence.start_date && planning.date !== formated && !alldates.some(date => format(date, 'dd/MM/yyyy', { locale: fr }) === planning.date)) {
                             db.prepare('DELETE FROM planning WHERE id = ?').run(planning.id);
                         }
                     });
+                    alldates.map(date => {
+                        const formatedDate = format(date, 'dd/MM/yyyy', { locale: fr });
+                        if (formatedDate !== nextDay(parsedStartDate) && !plannings.some(planning => planning.date === formatedDate)) {
+                            db.prepare('INSERT INTO planning (driver_id, date, client_name, start_time, return_time, note, destination, long_distance, recurrence_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)').run(0, formatedDate, last.client_name, last.start_time, last.return_time, last.note, last.destination, last.long_distance, recurrence.id);
+                        }
+                    });
                 } else if (frequency.length > 1) {
-                    console.log('long recurrence', frequency, recurrence.id);
                     let nextDates = [];
                     let allNextDates = [];
                     for (const day of frequency) {
-                        const potentialNextDate = nextWeekDay[day](parsedStartDate);
-                        allNextDates.push(potentialNextDate);
-                        const compareDate = potentialNextDate - now;
-                        const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
-                        nextDates.push({ [day]: difference });
+                        let currentDate = parsedStartDate;
+                        for (let i = 0; i < 4; i++) {
+                            const potentialNextDate = nextWeekDay[day](currentDate);
+                            allNextDates.push(potentialNextDate);
+                            const compareDate = potentialNextDate - now;
+                            const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
+                            nextDates.push({ [day]: difference });
+                            currentDate = potentialNextDate;
+                        }
                     }
                     nextDates.sort((a, b) => a[Object.keys(a)[0]] - b[Object.keys(b)[0]]);
                     const planningDate = plannings.map(planning => {
@@ -135,7 +167,6 @@ export function checkAll() {
                     })
                     if (!isPlanned) {
                         try {
-                            console.log('insert planning > 1 today', format(nextDate, 'dd/MM/yyyy', { locale: fr }));
                             db.prepare('INSERT INTO planning (driver_id, date, client_name, start_time, return_time, note, destination, long_distance, recurrence_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)').run(0, format(nextDate, 'dd/MM/yyyy', { locale: fr }), last.client_name, last.start_time, last.return_time, last.note, last.destination, last.long_distance, recurrence.id);
                         } catch (err) {
                             console.error('Error while adding planning: ', err);
@@ -143,7 +174,6 @@ export function checkAll() {
                                 db.prepare('DELETE FROM recurrence WHERE id = ?').run(recurrence.id);
                             }
                         }
-                        console.log('insert planning > 1 not today', format(nextDate, 'dd/MM/yyyy', { locale: fr }));
                     }
                     const data = excludeDays?.find(exclude => exclude.recurrence_id === recurrence.id);
                     const exdate = JSON.parse(data?.date || "[]");
@@ -156,6 +186,68 @@ export function checkAll() {
                     for (const line of plannings) {
                         if (parse(line.date, 'dd/MM/yyyy', new Date(), { locale: fr }) > parse(recurrence.start_date, 'dd/MM/yyyy', new Date(), { locale: fr }) && !formatedAllNextDate.includes(line.date)) {
                             db.prepare('DELETE FROM planning WHERE id = ?').run(line.id);
+                        }
+                    }
+                } else if (!isToday && startDateParsed < now && nextDate < now) {
+                    let find = false;
+                    const maxIterations = 360;
+                    let iterationCount = 0;
+                    if (recurrence.length === 1) {
+                        let nextDay = nextDate;
+                        let startDay;
+                        while (find === false && maxIterations > iterationCount) {
+                            const searchingNextDate = nextWeekDay[frequency[0]](nextDay);
+                            const compareDate = searchingNextDate - now;
+                            const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
+                            if (difference >= 0) {
+                                find = true;
+                                if (difference === 0) {
+                                    startDay = searchingNextDate
+                                }
+                                nextDay = nextWeekDay[frequency[0]](searchingNextDate);
+                            } else {
+                                nextDay = searchingNextDate;
+                            }
+                            iterationCount++;
+                        }
+                        if (startDay !== undefined) {
+                            db.prepare('UPDATE recurrence SET start_date = ?, next_day = ? WHERE id = ?').run(format(startDay, 'dd/MM/yyyy', { locale: fr }), format(nextDay, 'dd/MM/yyyy', { locale: fr }), recurrence.id);
+                        } else {
+                            db.prepare('UPDATE recurrence SET start_date = ?, next_day = ? WHERE id = ?').run(format(nextDate, 'dd/MM/yyyy', { locale: fr }), format(nextDay, 'dd/MM/yyyy', { locale: fr }), recurrence.id);
+                        }
+                    } else if (recurrence.length > 1) {
+                        let nextDay = nextDate;
+                        let startDay;
+                        let nextDateToSearch;
+                        while (find === false && maxIterations > iterationCount) {
+                            let allDifferences = [];
+                            for (const day of frequency) {
+                                const searchingNextDate = nextWeekDay[day](nextDay);
+                                const compareDate = searchingNextDate - now;
+                                const difference = Math.round(compareDate / (1000 * 60 * 60 * 24));
+                                allDifferences.push({ [day]: difference });
+                            }
+                            allDifferences.sort((a, b) => a[Object.keys(a)[0]] - b[Object.keys(b)[0]]);
+                            const minKey = Object.keys(allDifferences[0])[0];
+                            const searchingNextDate = nextWeekDay[minKey](nextDay);
+                            if (allDifferences[0] >= 0) {
+                                find = true;
+                                if (difference === 0) {
+                                    startDay = searchingNextDate
+                                }
+                                nextDay = nextWeekDay[frequency[0]](searchingNextDate);
+                            } else {
+                                nextDay = searchingNextDate;
+                            }
+                            if (iterationCount >= maxIterations) {
+                                console.error('Max iterations reached while searching for the next date.');
+                            }
+                            iterationCount++;
+                        }
+                        if (startDay !== undefined) {
+                            db.prepare('UPDATE recurrence SET start_date = ?, next_day = ? WHERE id = ?').run(format(startDay, 'dd/MM/yyyy', { locale: fr }), format(nextDay, 'dd/MM/yyyy', { locale: fr }), recurrence.id);
+                        } else {
+                            db.prepare('UPDATE recurrence SET start_date = ?, next_day = ? WHERE id = ?').run(format(nextDate, 'dd/MM/yyyy', { locale: fr }), format(nextDay, 'dd/MM/yyyy', { locale: fr }), recurrence.id);
                         }
                     }
                 }
